+++ factstore-avro/src/test/kotlin/org/factstore/avro/AvroFactStoreTest.kt
@@ -7,6 +7,7 @@ import kotlinx.serialization.Contextual
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import org.factstore.core.*
+import org.factstore.foundationdb.buildFdbFactStore
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
