+++ factstore-foundationdb/src/main/kotlin/org/factstore/foundationdb/FdbFact.kt
@@ -1,6 +1,7 @@
-package org.factstore.core
+package org.factstore.foundationdb
 
 import com.apple.foundationdb.tuple.Tuple
+import org.factstore.core.Fact
 
 data class FdbFact(
     val fact: Fact,
